Fix the `get_name` function
Previously, the `get_name` function passed a file extension to the
rstrip method. However, this method doesn't strip *all* characters
passed to it; it strips *any* characters passed to it. For example:

>>> name = "input.feather"
>>> name.rstrip(".feather")
'inpu'

In this example, the file extension contains the letter "t", so rstrip
strips the letter "t" from "input".

Another example:

>>> name = "input.csv"
>>> name.rstrip(".csv")
'input'

In this example, the file extension has no letters in common with
"input", so rstrip doesn't strip any letters from "input".