Replace count and percentage with empty
The count and percentage of missing values for each column needed to be
rounded and suppressed. However, what we really want to know is whether
each column is empty; this doesn't need to be rounded and suppressed.
Consequently, we replace the count and percentage with empty. This makes
for cleaner code and cleaner separation of a dataset report: the Summary
section is "safe"; the Columns section is "unsafe", but two disclosure
controls (rounding, suppression) have been applied.